获取数据库动态数据

+++ hosptial_entry/src/main/java/com/lck/repository/UserRepository.java
@@ -6,8 +6,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import org.springframework.stereotype.Component;
 
+import javax.persistence.criteria.CriteriaBuilder;
+
 @Component
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
     @Query(value = "select id,user_name,password,grade from t_user u where u.user_name =?1",nativeQuery=true)
     User findByUserName(String userName);
 }
